+++ Changelog.md
@@ -3,6 +3,7 @@
 ## Version (NEXT)
 * added domain model elements
 * updated banking model
+* add arc42 templates
 
 ## Version 0.41.0
 * added :instance-of relation