+++ Changelog.md
@@ -4,6 +4,7 @@
 * added domain model elements
 * updated banking model
 * add arc42 templates
+* updated dependencies
 
 ## Version 0.41.0
 * added :instance-of relation