+++ Changelog.md
@@ -3,8 +3,11 @@
 ## Version (NEXT)
 * adding views for traceability (e.g. instanciation/implementation views)
 * added domain model elements for domain-driven design
-  * :domain, :bounded-context, :aggregate, :domain-event, :command and :policy nodes
-  * :triggers relation 
+  * :domain, :bounded-context, :aggregate, :entity :value-object, :domain-event, :command and :policy nodes
+  * :triggers relation
+* added process model elements
+  * :control, :goal, :regulation and :test nodes
+  * :control-for, :goal-for, :regulation-for and :test-for relations
 * added :container-instance node to deployment model
 * added loading of template library namespaces from templates
 * added template library